Split tab nav items into a separate file
